fix 8xy4 and 9xy0, add window clearance on new render

+++ Sharp8.Core/Helpers.cs
@@ -2,5 +2,5 @@ namespace Sharp8;
 
 public abstract class Helpers
 {
-    public static int MsbByteMask {get;} = 0x80;
+    public static byte MsbByteMask { get; } = 0x80;
 }
